+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/api/platform/configs/fabric/ConfigBuilderImpl.java
@@ -68,6 +68,7 @@ public class ConfigBuilderImpl extends ConfigBuilder {
         var tooltipKey = this.tooltipKey(name);
         if (this.translations.containsKey(tooltipKey)) {
             config.setDescriptionKey(tooltipKey);
+            config.setRawDescription(this.translations.get(tooltipKey));
         }
 
         this.categoryStack.peek().addEntry(config);
@@ -164,8 +165,4 @@ public class ConfigBuilderImpl extends ConfigBuilder {
         return this;
     }
 
-    @Override
-    public ConfigBuilder comment(String comment) {
-        return super.comment(comment);
-    }
 }