+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/api/platform/configs/fabric/values/ColorConfigValue.java
@@ -40,4 +40,8 @@ public class ColorConfigValue extends IntConfigValue {
     }
 
 
+    @Override
+    public String getExtraInfo() {
+        return "Accepted format: Hexadecimal color code (e.g., #RRGGBB or #AARRGGBB)";
+    }
 }