+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/api/platform/configs/fabric/values/DoubleConfigValue.java
@@ -52,4 +52,9 @@ public class DoubleConfigValue extends ConfigValue<Double> {
     public Double getMin() {
         return min;
     }
+
+    @Override
+    public String getExtraInfo() {
+        return "Accepted range: " + min + " to " + max;
+    }
 }