+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/api/platform/configs/fabric/values/IntConfigValue.java
@@ -52,4 +52,9 @@ public class IntConfigValue extends ConfigValue<Integer> {
     public Integer getMin() {
         return min;
     }
+
+    @Override
+    public String getExtraInfo() {
+        return "Accepted range: " + min + " to " + max;
+    }
 }