+++ neoforge/src/main/java/net/mehvahdjukaar/moonlight/core/misc/neoforge/FilteredResManagerImpl.java
@@ -1,0 +1,10 @@
+package net.mehvahdjukaar.moonlight.core.misc.neoforge;
+
+import net.minecraft.server.packs.PackResources;
+
+public class FilteredResManagerImpl {
+    public static boolean isModResourcePack(PackResources pack) {
+        String packId = pack.packId();
+        return packId.equals("mod_resources") || packId.startsWith("mod/") || packId.startsWith("generated");
+    }
+}